Simplify project structure
Instead of 4 independent gradle projects we now have one project with 4
modules.  This cuts down on the build system overhead and makes it
easier to bootstrap the project in Android Studio, because you only need
to open one project instead of 4.  As a side benefit, the apk's have
more descriptive names than "app-debug.apk".