fix positioning of face points for downscaled images
The face points returned from the API are relative to the dimensions of the
original image, but the Frame object passed to onImageResults contains a
BitmapFrame which may have been scaled down by the SDK (in addition
to the original bitmap).

The ImageDetectorDemo app was positioning the face points on the screen
by scaling them to reflect the relative dimensions of the scaled down image
in the Frame and the dimensions of the displayed image on screen. I fixed it
by having it use the dimensions of the original image instead of the
dimensions of the scaled down image.