address timestamp precision issue in FrameDetectorDemo
FrameDetector Demo was sending epoch-based frame timestamps.  Due to the precision
limit of floats in Java and the large values of epoch-based  timestamps, this resulted in
imprecise values being delivered to the detector.  This commit changes the approach to
use first-frame-relative timestamps instead of epoch-relative values.

also updated gradle plugin version