3.2 SDK updates & other build misc
add UI for displaying brightness value in CameraDetectorDemo,
FrameDetectorDemo, and VideoDetectorDemo.

consolidate gradle stuff which is common to all modules in
module-defaults.gradle.  This simplified the module build.gradle files
considerably.

update affdexsdk dependencies for all sample apps to use dynamic versioning
with value of "3.+", which will cause them to pick up the highest-numbered
SDK release in the 3.x series.  This way we don't necessarily need to update the
samples when we do a point release.

update gradle plugin to 2.2.2

set all versionCode values to 1 (normally a Bad Thing if you're app is
published to the Play store, but these aren't, so simpler to just have it
this way)

set APK naming style to appname-version-build#